- 侧边栏提取为组件，简化App.vue。 - 添加登录对话框的回车登录功能。
